refactor: logging
