+++ smarter/smarter/lib/django/waffle/switches.py
@@ -1,90 +1,11 @@
 """
-smarter.lib.django.waffle
-------------------------------
-
-Enhanced, managed Django-waffle wrapper with short-lived Redis-based caching
-and database readiness checks.
-
-Features:
-
-- Caching: Integrates short-lived Redis-based caching to optimize feature flag (switch) checks.
-- Database Readiness Handling: Implements safeguards to prevent errors when the database is not ready.
-- Feature Flag Management: Centralized mechanism to check if a feature flag (switch) is active.
-- Custom Django Admin: Customized Django Admin class for managing waffle switches.
-- Fixed Set of Switches: Defines a fixed set of waffle switches for the Smarter API.
-
-.. important::
-
-    These are managed feature flags; add any new switches to the SmarterWaffleSwitches class. These switches
-    are verified duing deployments to ensure that they exist in the database. Missing switches are
-    automatically created with a default inactive state.
-
-.. important::
-
-    django-waffle relies on database tables as well as Redis for storing and caching feature flags. If the database is not ready, waffle switch values
-    will default to inactive (False) to prevent application errors.
-
-Example:
-    .. code-block:: python
-
-        from smarter.lib.django.waffle import switch_is_active
-
-        if switch_is_active(SmarterWaffleSwitches.ACCOUNT_LOGGING):
-            print("API logging is enabled.")
-
-Dependencies:
-
-- `django-waffle <https://waffle.readthedocs.io/en/stable/>`_
-- `django-redis <https://django-redis.readthedocs.io/en/stable/>`_
-- `Redis <https://redis.io/documentation>`_
-- `Django <https://docs.djangoproject.com/en/stable/>`_
+Module: smarter.lib.django.waffle.switches
 """
 
-import logging
 from dataclasses import dataclass
-
-import waffle as waffle_orig
-from django.apps import apps
-from django.core.exceptions import AppRegistryNotReady, SynchronousOnlyOperation
-from django.db import connections
-from django.db.utils import OperationalError, ProgrammingError
-from waffle.admin import SwitchAdmin
-
-# Also catch MySQLdb.OperationalError for lower-level DB errors
-try:
-    import MySQLdb
-
-    MySQLdbOperationalError = MySQLdb.OperationalError
-except ImportError:
-    MySQLdbOperationalError = None
-
-from smarter.common.helpers.console_helpers import formatted_text
-
-logger = logging.getLogger(__name__)
-prefix = f"{formatted_text(__name__)}.switch_is_active()"
-cache_prefix = __name__
 
 
 # pylint: disable=C0115
-class DbState:
-    ready = False
-
-
-db_state = DbState()
-
-
-class SmarterSwitchAdmin(SwitchAdmin):
-    """
-    Customized Django Admin console for managing Waffle switches.
-    This class restricts access to the module to superusers only.
-    """
-
-    ordering = ("name",)
-
-    def has_module_permission(self, request):
-        return hasattr(request, "user") and hasattr(request.user, "is_superuser") and request.user.is_superuser  # type: ignore[return-value]
-
-
 @dataclass(frozen=True)
 class SmarterWaffleSwitch:
     name: str
@@ -432,82 +353,4 @@ smarter_waffle_switches = SmarterWaffleSwitches()
 Singleton instance of SmarterWaffleSwitches to be used throughout the codebase.
 """
 
-
-def is_database_ready(alias="default"):
-    """
-    Check if the database is ready by verifying the connection and
-    the existence of the waffle_switch table.
-
-    :param alias: The database alias to check.
-    :return: True if the database is ready, False otherwise.
-    """
-
-    if db_state.ready:
-        return True
-    try:
-        # Ensure the connection is usable. ie the DB server is up
-        connection = connections[alias]
-        connection.ensure_connection()
-        # Check if the waffle_switch table exists
-        if "waffle_switch" not in connection.introspection.table_names():
-            return False
-        db_state.ready = True
-        return db_state.ready
-    except (OperationalError, SynchronousOnlyOperation):
-        # we'll get SynchronousOnlyOperation whenever an ASGI asynchronous
-        # context tries to interact with the database. Main culprit is middleware.
-        return False
-    except ProgrammingError:
-        return False
-
-
-def switch_is_active(switch_name: str) -> bool:
-    """
-    Check if a Waffle switch is active, with caching and database readiness checks.
-
-    .. important::
-
-        This is the preferred method for checking Waffle switches in the Smarter codebase.
-        It includes caching to optimize performance and handles database readiness to prevent errors.
-
-    Example:
-        .. code-block:: python
-
-            from smarter.lib.django.waffle import SmarterWaffleSwitches, switch_is_active
-
-            if switch_is_active(SmarterWaffleSwitches.API_LOGGING):
-                print("API logging is enabled.")
-
-    :param switch_name: The name of the Waffle switch to check.
-    :return: True if the switch is active, False otherwise.
-    :rtype: bool
-    """
-    # Prevent model access before Django app registry is ready
-    if not apps.ready:
-        logger.warning("%s App registry not ready, assuming switch %s is inactive.", prefix, switch_name)
-        return False
-
-    try:
-        if not is_database_ready():
-            logger.debug("%s Database not ready, assuming switch %s is inactive.", prefix, switch_name)
-            return False
-    # pylint: disable=broad-except
-    except Exception as e:
-        logger.warning("%s Error checking database readiness: %s", prefix, e, exc_info=True)
-        return False
-    if not isinstance(switch_name, str):
-        logger.error("%s switch_name must be a string, got %s", prefix, type(switch_name).__name__)
-        return False
-    if switch_name not in smarter_waffle_switches.all:
-        logger.error("%s switch_name '%s' is not a valid SmarterWaffleSwitches attribute", prefix, switch_name)
-        return False
-    db_exceptions = tuple(
-        t for t in (OperationalError, ProgrammingError, MySQLdbOperationalError) if t is not None
-    ) or (Exception,)
-    try:
-        return waffle_orig.switch_is_active(switch_name)
-    except (*db_exceptions, AppRegistryNotReady) as e:
-        logger.error(
-            "%s Database not ready, App Registry not ready, or switch does not exist: %s", prefix, e, exc_info=True
-        )
-        return False
+__all__ = ["SmarterWaffleSwitches", "smarter_waffle_switches"]
